move components files to separate folders
