+++ src/App.js
@@ -1,4 +1,4 @@
-import Scorad from "./Scorad";
+import Scorad from "./components/Scorad/Scorad";
 
 function App() {
   return (
