+++ src/components/Symptoms/Symptoms.js
